% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_models.R
\name{arfima}
\alias{arfima}
\title{ARFIMA model fitting and updates}
\usage{
arfima(y, model = NULL, lambda = NULL, stepwise = FALSE, ...)
}
\arguments{
\item{y}{Univariate Time Series}

\item{model}{\code{tsm} object, used to refit model}

\item{lambda}{optional, if null ignored otherwise treated as the Box-Cox
parameter}

\item{stepwise}{logical. should arma coefficients be fit by stepwise selection
or by checking all possible models? FALSE indicates all possible models.}

\item{...}{additional argument for model function}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{arfima}} to
allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package = 'datasets')
fit <- arfima(AirPassengers)
fit2 <- arfima(AirPassengers, lambda = 0)
}
\seealso{
\code{\link[forecast]{arfima}}

Other Time Series Models: \code{\link{arima}},
  \code{\link{bats}}, \code{\link{ets}}
}

