% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{arima}
\alias{arima}
\title{ARIMA model fitting and updates}
\usage{
arima(y, ..., model.name = "arima")
}
\arguments{
\item{y}{Univariate Time Series}

\item{...}{additional argument for model function}

\item{model.name}{name of model specialization}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{Arima}} and
\code{\link[forecast]{auto.arima}} to allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
fit <- arima(AirPassengers)
fit2 <- arima(AirPassengers, D=1)
fit3 <- arima(AirPassengers, lambda=0)
}
\seealso{
\code{\link[forecast]{Arima}} and \code{\link[forecast]{auto.arima}}
}

