% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.tests.R
\name{autocorr.lags}
\alias{autocorr.lags}
\title{Identification of lag threshold for Autocorrelation analysis}
\usage{
autocorr.lags(y, alpha = 0.05)
}
\arguments{
\item{y}{univariate time series}

\item{alpha}{numeric, determines significance threshold for p-value analysis}
}
\value{
integer
}
\description{
Identification of lag threshold for Autocorrelation analysis
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
autocorr.lags(AirPassengers)
}
\seealso{
Other Autocorrelation Tests: \code{\link{durbin.watson.test}},
  \code{\link{ljung.box.test}},
  \code{\link{mcleod.li.test}}, \code{\link{print.act}}
}

