% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.tests.R
\name{autocorr.signif.lags}
\alias{autocorr.signif.lags}
\title{Identify lags threshold for autocorrelation analysis}
\usage{
autocorr.signif.lags(y, alpha = 0.05)
}
\arguments{
\item{y}{univariate time series}

\item{alpha}{numeric, determines significance threshold for p-value analysis}
}
\value{
integer
}
\description{
Identify lags threshold for autocorrelation analysis
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
autocorr.signif.lags(AirPassengers)
}

