% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.ts}
\alias{autoplot.ts}
\title{Automatic plotting for time series objects}
\usage{
autoplot.ts(object, title = NULL, ...)
}
\arguments{
\item{object}{a \code{ts} or \code{mts} object}

\item{title}{character. If provided, is passed as an argument to
\code{\link[ggplot2]{ggtitle}}}

\item{...}{other arguments}
}
\description{
Automatic plotting for time series objects
}
\examples{
library(ggplot2)
library(forecastR)

ap <- ts.split(AirPassengers, as.list = FALSE)
autoplot(ap, title = "Air Passengers Subsetting")
}

