% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{baseline.forecast}
\alias{baseline.forecast}
\title{Baseline / Null Hypothesis Forecasts}
\usage{
baseline.forecast(y, h = 18L, lambda = optimize.lambda(y), ...)
}
\arguments{
\item{y}{\code{ts} univariate time series}

\item{h}{integer. number of periods to forecast}

\item{lambda}{numeric. parameter for Box-Cox transformation}

\item{...}{additional arguments}
}
\description{
Baseline / Null Hypothesis Forecasts
}

