% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_models.R
\name{bats}
\alias{bats}
\title{BATS model fitting and updates}
\usage{
bats(y, model = NULL, lambda = NULL, ...)
}
\arguments{
\item{y}{Univariate Time Series}

\item{model}{previously fitted model object.}

\item{lambda}{not used, in place to ensure uniform API}

\item{...}{additional argument for model function}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{bats}} to
allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package = 'datasets')
fit <- bats(AirPassengers)
fit2 <- bats(AirPassengers, lambda = 0)
}
\seealso{
\code{\link[forecast]{bats}}

Other Time Series Models: \code{\link{arfima}},
  \code{\link{arima}}, \code{\link{ets}}
}

