% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{bootforecast}
\alias{bootforecast}
\alias{bootforecast.tsm}
\title{Bootstrap Forecast for model}
\usage{
bootforecast(model, y, bootreps = 100, split = 0.2, ...)

\method{bootforecast}{tsm}(model, y, bootreps = 100, split = 0.2, ...)
}
\arguments{
\item{model}{\code{tsm} object fit against sample of data}

\item{y}{\code{ts} Univariate time series to be bootstrapped}

\item{bootreps}{integer. number of bootstrap replicates}

\item{split}{numeric. value passed to \code{\link[forecastR]{ts.split}}}

\item{...}{additional arguments}
}
\value{
\code{bootforecast} object
}
\description{
Bootstrap Forecast for model

\code{tsm} Bootstrap forecasts
}
\section{Methods (by class)}{
\itemize{
\item \code{tsm}: \code{tsm}
}}
\seealso{
\code{\link[meboot]{meboot}} \code{\link[forecastR]{ts.split}}
}

