% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanse.R
\name{cleanse.attributes}
\alias{cleanse.attributes}
\title{Extract cleanse attributes from time series}
\usage{
cleanse.attributes(x, ...)
}
\arguments{
\item{x}{univariate time series}

\item{...}{additional arguments}
}
\description{
Extract cleanse attributes from time series
}

