% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanse.R
\name{cleanse}
\alias{cleanse}
\title{Cleanse Time Series in preparation for fitting}
\usage{
cleanse(x, ...)
}
\arguments{
\item{x}{\code{numeric}}

\item{...}{additional arguments}
}
\description{
Cleanse Time Series in preparation for fitting
}

