% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsm.R
\name{coef.tsm}
\alias{coef.tsm}
\title{Extract Model Coefficients}
\usage{
\method{coef}{tsm}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is meaningful.}

\item{...}{other arguments}
}
\description{
Extract Model Coefficients
}
\seealso{
\code{\link[stats]{coef}}
}

