% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.tests.R
\name{durbin.watson.test}
\alias{durbin.watson.test}
\title{Durbin-Watson test}
\usage{
durbin.watson.test(y, alpha = 0.05, max.lag = autocorr.signif.lags(y,
  alpha))
}
\arguments{
\item{y}{univariate time series}

\item{alpha}{numeric, determines significance threshold for p-value analysis}

\item{max.lag}{integer. If undefined, calculates an appropriate value}
}
\value{
\code{'DurbinWatsonTest'} object
}
\description{
Durbin-Watson test
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
aa <- arima(AirPassengers)
durbin.watson.test(resid(aa))
aa2 <- arima(AirPassengers, lambda=0)
durbin.watson.test(resid(aa2))

#check for GARCH residuals
durbin.watson.test(resid(aa)^2)
durbin.watson.test(resid(aa2)^2)
}
\seealso{
\code{\link[car]{durbinWatsonTest}}
}

