% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model_fit.R
\name{final.forecast.selection}
\alias{final.forecast.selection}
\title{Final forecast selection}
\usage{
final.forecast.selection(y, ts.mixture, split = 0.1, ...)
}
\arguments{
\item{y}{time series vector}

\item{ts.mixture}{object of class 'tsm'}

\item{split}{fraction of records to reserve for our of sample cross validation}

\item{...}{other keyword arguments, not currently used}
}
\description{
Final forecast selection
}

