% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsm.R
\name{fitted.tsm}
\alias{fitted.tsm}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{tsm}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model fitted values is meaningful.}

\item{...}{other arguments}
}
\description{
Extract Model Fitted Values
}
\seealso{
\code{\link[stats]{fitted}}
}

