% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast \code{tsm} object model}
\usage{
forecast(object, h = 18L, ...)
}
\arguments{
\item{object}{\code{\link[forecastR]{tsm}} fitted model object.}

\item{h}{integer. number of out of sample forecast periods}

\item{...}{additional argument for model function}
}
\value{
\code{\link[forecast]{forecast}} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{forecast}} to
allow unified interface.
}
\examples{
library(ggplot2)
library(forecastR)

data('AirPassengers', package='datasets')
fit <- arima(AirPassengers)
fcst <- forecast(fit)

library(forecast)
(autoplot(fcst))
}
\seealso{
\code{\link[forecast]{forecast}}
}

