% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{forecastR-package}
\alias{forecastR}
\alias{forecastR-package}
\title{forecastR}
\description{
This package relies heavily on \code{\link{forecast}} and \code{\link{opera}}
to help forecasters select and build high-quality forecasts.
}
\author{
Robert Smith <rob_smith@goodyear.com>
}
\keyword{package}

