% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model_fit.R
\name{generate.forecast}
\alias{generate.forecast}
\title{Generate forecast}
\usage{
generate.forecast(y, boot.reps = NULL, oos.h = 18L, split = 0.2,
  alpha = 0.05, ...)
}
\arguments{
\item{y}{time series vector}

\item{boot.reps}{number of maximum entropy bootstrap replicates. If \code{NULL}, then no bootstrapping is applied.}

\item{oos.h}{integer. number of out of sample forecast steps}

\item{split}{fraction of observations from \code{y} which should be reserved for cross validation}

\item{alpha}{significance threshold}

\item{...}{other keyword arguments, not currently used}
}
\description{
Generate forecast
}

