% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanse.R
\name{inv.cleanse}
\alias{inv.cleanse}
\title{Invert cleansed time series}
\usage{
inv.cleanse(x, cleansed.x = NULL, ...)
}
\arguments{
\item{x}{univariate time series}

\item{cleansed.x}{cleansed attr}

\item{...}{additional arguments}
}
\description{
Invert cleansed time series
}

