% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.tests.R
\name{ljung.box.test}
\alias{ljung.box.test}
\title{Ljung-Box Test}
\usage{
ljung.box.test(y, alpha = 0.05, max.lag = autocorr.signif.lags(y, alpha))
}
\arguments{
\item{y}{univariate time series}

\item{alpha}{numeric, determines significance threshold for p-value analysis}

\item{max.lag}{integer. If undefined, calculates an appropriate value}
}
\value{
\code{'LjungBoxTest'} object
}
\description{
Ljung-Box Test
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
aa <- arima(AirPassengers)
ljung.box.test(resid(aa))
aa2 <- arima(AirPassengers, lambda=0)
ljung.box.test(resid(aa2))

#check for GARCH residuals
ljung.box.test(resid(aa)^2)
ljung.box.test(resid(aa2)^2)
}
\seealso{
\code{\link[stats]{Box.test}}
}

