% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.tests.R
\name{mcleod.li.test}
\alias{mcleod.li.test}
\title{McLeod - Li Test}
\usage{
mcleod.li.test(y, alpha = 0.05, max.lag = autocorr.signif.lags(y, alpha))
}
\arguments{
\item{y}{univariate time series}

\item{alpha}{numeric, determines significance threshold for p-value analysis}

\item{max.lag}{integer. If undefined, calculates an appropriate value}
}
\value{
\code{'McLeodLiTest'} object
}
\description{
McLeod - Li Test
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
aa <- arima(AirPassengers)
mcleod.li.test(aa)
aa2 <- arima(AirPassengers, lambda=0)
mcleod.li.test(aa2)

#check for GARCH residuals
mcleod.li.test(resid(aa)^2)
mcleod.li.test(resid(aa2)^2)
}
\seealso{
\code{\link[TSA]{McLeod.Li.test}}
}

