% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsm.R
\name{model<-}
\alias{model<-}
\title{Time Series Model Assignnment}
\usage{
model(object) <- value
}
\arguments{
\item{object}{object which is to store a time series model}

\item{value}{time series model to be stored}
}
\description{
Time Series Model Assignnment
}

