% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsm.R
\name{model}
\alias{model}
\title{Extract a model from an object}
\usage{
model(object, ...)
}
\arguments{
\item{object}{object to be evaluated}

\item{...}{additional arguments for class specific evaluation}
}
\value{
model object
}
\description{
Extract a model from an object
}

