% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{multiforecast}
\alias{multiforecast}
\title{Forecast for \code{tsm.multi} objects}
\usage{
multiforecast(object.list, h = 18L, y = NULL, ...)
}
\arguments{
\item{object.list}{list of \code{tsm} objects}

\item{h}{integer. length of forecast output}

\item{y}{\code{ts}. Optional univariate time series, if provided baseline
forecasts are also provided.}

\item{...}{additional arguments}
}
\value{
\code{mts} of forecast means w/ named columns
}
\description{
Forecast for \code{tsm.multi} objects
}

