% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nnetar}
\alias{nnetar}
\title{NNETAR model fitting and updates}
\usage{
nnetar(y, ..., model.name = "nnetar")
}
\arguments{
\item{y}{Univariate Time Series}

\item{...}{additional argument for model function}

\item{model.name}{name of model specialization}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{nnetar}} to
allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
fit <- nnetar(AirPassengers)
fit2 <- nnetar(AirPassengers, lambda=0)
}
\seealso{
\code{\link[forecast]{nnetar}}
}

