% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\name{package_options}
\alias{package_options}
\title{Set or get options for \code{forecastR} package}
\usage{
package_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for \code{forecastR} package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{models}}{
   \itemize{
     \item{\code{arima}}{ARIMA model will be considered, see also \code{\link[forecast]{Arima}} & \code{\link[forecast]{auto.arima}}}
     \item{\code{arfima}}{ARFIMA model will be considered, see also \code{\link[forecast]{arfima}}}
     \item{\code{bats}}{BATS models will be considered, see also \code{\link[forecast]{bats}}}
     \item{\code{ets}}{ETS model will be considered, see also \code{\link[forecast]{ets}}}
     \item{\code{nnetar}}{NNETAR model will be considered, see also \code{\link[forecast]{nnetar}}}
     \item{\code{stlm}}{STLM model will be considered, see also \code{\link[forecast]{stlm}}}
     \item{\code{tbats}}{TBATS model will be considered, see also \code{\link[forecast]{tbats}}}
     \item{\code{tslm}}{TSLM model will be considered, see also \code{\link[forecast]{tslm}}}
   }
 }
}
}

