% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.tests.R
\name{print.act}
\alias{print.act}
\title{Print classes of \code{act} (autocorrelation test)}
\usage{
\method{print}{act}(x, ...)
}
\arguments{
\item{x}{input object}

\item{...}{additional arguments for print function}
}
\description{
Print classes of \code{act} (autocorrelation test)
}

