% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsm.R
\name{residuals.tsm}
\alias{residuals.tsm}
\title{Extract Model Residuals}
\usage{
\method{residuals}{tsm}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model residuals is meaningful.}

\item{...}{other arguments}
}
\value{
Residuals extracted from object \code{object}.
}
\description{
Extract Model Residuals
}
\seealso{
\code{\link[stats]{residuals}}
}

