% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{short.ts.test}
\alias{short.ts.test}
\title{Tests if a time series is short}
\usage{
short.ts.test(y, freq.multiple = 2)
}
\arguments{
\item{y}{univariate time series}

\item{freq.multiple}{numeric. multiple of the frequency of x to be compared
against the length of x}
}
\description{
Used internally to determine the appropriate time series models to apply.
Short time series will be modeled as non-seasonal and will be fit against
simpler models.
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
print('Default Freq. Multiple:', package_options('short.ts.frequency.multiple'))
## [1] 2
short.ts.test(AirPassengers) #returns FALSE
short.ts.test(AirPassengers, 15.0) #returns TRUE
}

