% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsm.R
\name{summary.tsm}
\alias{summary.tsm}
\title{Object Summary for \code{tsm} objects}
\usage{
\method{summary}{tsm}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Object Summary for \code{tsm} objects
}
\seealso{
\code{\link[base]{summary}}
}

