% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{tbats}
\alias{tbats}
\title{TBATS model fitting and updates}
\usage{
tbats(y, ..., model.name = "tbats")
}
\arguments{
\item{y}{Univariate Time Series}

\item{...}{additional argument for model function}

\item{model.name}{name of model specialization}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{tbats}} to
allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
fit <- nnetar(AirPassengers)
fit2 <- nnetar(AirPassengers, lambda=0)
}
\seealso{
\code{\link[forecast]{tbats}}
}

