% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.ts}
\alias{tidy.ts}
\title{Automatic conversion of \code{ts} and \code{mts} objects to \code{data.frame}}
\usage{
tidy.ts(object, na.rm = TRUE, variable.name = "variable",
  value.name = "value", ...)
}
\arguments{
\item{object}{a \code{ts} or \code{mts} object}

\item{na.rm}{logical. If \code{TRUE}, when numeric values are coalesced,
missing observations are dropped, otherwise they are kept as \code{NA}}

\item{variable.name}{character. Name of column which indicates the source of
the values in `value.name`}

\item{value.name}{character. Name of the resulting column of values.}

\item{...}{other arguments}
}
\description{
Automatic conversion of \code{ts} and \code{mts} objects to \code{data.frame}
}
\examples{
library(broom)
library(forecastR)
ap <- ts.split(AirPassengers)
head(ap)

tidy.ap <- tidy(ap)
head(tidy.ap)
}

