% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy.ts.split}
\alias{tidy.ts.split}
\title{Tidy \code{ts.split} object into \code{data.frame}.}
\usage{
\method{tidy}{ts.split}(x, ...)
}
\arguments{
\item{x}{\code{ts.split} object}

\item{...}{arguments passed on to \code{as.data.frame}}
}
\value{
\code{\link{data.frame}} with `time` column and columns for
  each \code{\link[stats]{ts}} object. If \code{\link[stats]{ts}} inherits
  from \code{mts}, \code{\link{matrix}} or is a
  \code{\link{list}} of \code{\link[stats]{ts}} objects, it is coerced
  to a \code{mts} before using the native call to
  \code{\link{as.data.frame.ts}}.
}
\description{
Tidy \code{ts.split} object into \code{data.frame}.
}
\examples{
library(forecastR)
library(broom)
x <- ts(1:100, freq=12)
x.split <- ts.split(x)
x.tidy <- tidy(x.split)
is.data.frame(x.tidy)
head(x.tidy)
}
\seealso{
\code{\link[broom]{tidy}} \code{\link{as.data.frame.ts}}
  \code{\link{as.data.frame.list}}
}

