% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_characteristics.R
\name{ts.characteristics}
\alias{ts.characteristics}
\title{Time Series Characteristics}
\usage{
ts.characteristics(y)
}
\arguments{
\item{y}{time series vector}
}
\description{
Time Series Characteristics
}

