% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model_fit.R
\name{ts.mixture}
\alias{ts.mixture}
\title{Mixture Model fit}
\usage{
ts.mixture(y, tsm.multi, split = 0.2, oos.h = 18L, alpha = 0.05,
  boot.reps = 25, ...)
}
\arguments{
\item{y}{\code{\link[stats]{ts}} object. univariate time series}

\item{tsm.multi}{\code{tsm.multi} object returned from \code{\link[forecastR]{ts.multimodel.fit}}}

\item{split}{argument passed to \code{ts.split}}

\item{oos.h}{integer. Out of sample forecast steps to be provided in the
output.}

\item{alpha}{significance level for comparative testing}

\item{boot.reps}{integer. number of bootstrap replicates to fit against}

\item{...}{extra arguments for `ts.model.type`}
}
\description{
Mixture Model fit
}
\examples{
library(ggplot2)
library(forecastR)
data("AirPassengers", package="datasets")

ap <- window(AirPassengers, end=c(1959,12))
ap.oos <- window(AirPassengers, start=c(1960,1))

split.fctr <- 24
tsm <- ts.multimodel.fit(ap, split=split.fctr)

mx <- ts.mixture(ap, tsm, split=split.fctr, oos.h = 12)

vals <- cbind(AirPassengers, mx$mixture.model$response, mx$out.of.sample.experts)
colnames(vals) <- c("Data", "Mixture Model", colnames(mx$out.of.sample.experts))
autoplot(window(vals[,1:2], start=1959))

fcst <- cbind(mx$mixture.model$response, mx$out.of.sample.experts)
fcst.err <- fcst - ap.oos
colnames(fcst.err) <- c("Mixture Model", colnames(mx$out.of.sample.experts))
autoplot(abs(fcst.err))
cum.fcst.err <- ts(do.call(cbind, lapply(abs(fcst.err), cumsum)))
tsp(cum.fcst.err) <- tsp(fcst.err)
autoplot(cum.fcst.err)
}

