% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofit.R
\name{ts.model.fit}
\alias{ts.model.autofit}
\alias{ts.model.fit}
\title{Generic Time Series Model Fit for Package Models}
\usage{
ts.model.fit(y, ts.model.type = package_options("autofit.models"), ...)

ts.model.autofit(y, ts.model.type = package_options("autofit.models"),
  lambda = list(NULL, 0, forecast::BoxCox.lambda(y, "loglik")),
  alpha = 0.05, ...)
}
\arguments{
\item{y}{\code{ts}. Univariate Time Series}

\item{ts.model.type}{character. Name of model function from package.}

\item{...}{additional keyword arguments provided to `ts.model.type` function.}

\item{lambda}{list of values to pass to \code{lambda} values}

\item{alpha}{numeric. significance threshold for \code{\link[forecast]{dm.test}}
which determines if transformation makes for an improved model}
}
\value{
\code{tsm} object, with subspecialization determined by `ts.model.type`
argument.
}
\description{
Generic Time Series Model Fit for Package Models

Autofit Time Series models attempting multiple standardized transformations
}
\section{Functions}{
\itemize{
\item \code{ts.model.autofit}: Autofit
}}
\examples{
library(ggplot2)
library(forecastR)
data('AirPassengers', package='datasets')

ap.split <- ts.split(AirPassengers)

mf <- ts.model.fit(ap.split$in.sample, ts.model.type='arima')
mf.fcst <- forecast(mf, length(ap.split$out.of.sample))
summary(mf)
coef(mf)

autoplot(resid(mf))

vals <- cbind(Data=ap.split$data, Forecast.Mean=mf.fcst$mean)
autoplot(window(vals, start=c(1958,1)))
}
\seealso{
\code{\link[forecastR]{arima}} \code{\link[forecastR]{arfima}}
  \code{\link[forecastR]{bats}} \code{\link[forecastR]{ets}}
  \code{\link[forecastR]{nnetar}} \code{\link[forecastR]{tbats}}
  \code{\link[forecast]{auto.arima}} \code{\link[forecast]{Arima}}
  \code{\link[forecast]{arfima}} \code{\link[forecast]{bats}}
  \code{\link[forecast]{ets}} \code{\link[forecast]{nnetar}}
  \code{\link[forecast]{tbats}}
}

