% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofit.R
\name{ts.model}
\alias{ts.model}
\title{Compare and analyze models}
\usage{
ts.model(y, split = 0.2, oos.h = 18L, alpha = 0.05,
  ts.model.type = c("arima", "sarima", "arfima", "bats", "ets",
  "ets.multiplicative", "nnetar", "nnetar.w.decay", "tbats"), ...)
}
\arguments{
\item{y}{\code{\link[stats]{ts}} object. univariate time series}

\item{split}{argument passed to \code{\link[forecastR]{ts.split}}}

\item{oos.h}{integer. Out of sample forecast steps to be provided in the
output.}

\item{alpha}{significance level for comparative testing}

\item{ts.model.type}{character. name of function from package}

\item{...}{extra arguments for `ts.model.type`}
}
\value{
\code{model.fit}
}
\description{
Compare and analyze models
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
mf <- ts.model(AirPassengers, ts.model.type='arima')
}

