% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofit.R
\name{ts.model.refit}
\alias{ts.model.refit}
\title{Update Generic Time Series Model Fit for Package Models}
\usage{
ts.model.refit(y, model, ...)
}
\arguments{
\item{y}{\code{ts}. Univariate Time Series}

\item{model}{\code{tsm} object. Model which was fit in call to
\code{\link[forecastR]{ts.model.fit}}}

\item{...}{additional keyword arguments provided to `ts.model.type` function.}
}
\value{
\code{tsm} object, with subspecialization determined by `ts.model.type`
argument.
}
\description{
Update Generic Time Series Model Fit for Package Models
}
\examples{
library(ggplot2)
library(forecastR)
data('AirPassengers', package='datasets')

ap.split <- ts.split(AirPassengers)

mf <- ts.model.fit(ap.split$in.sample, ts.model.type='arima')
mf.fcst <- forecast(mf, length(ap.split$out.of.sample))
summary(mf)

vals <- cbind(Data=ap.split$data, Forecast.Mean=mf.fcst$mean)
autoplot(window(vals, start=c(1958,1)))

mf2 <- ts.model.refit(ap.split$data, mf)
mf2.fcst <- forecast(mf2, 24)
summary(mf2)

vals <- cbind(Data=ap.split$data, In.Sample.Fcst=mf.fcst$mean, Out.Of.Sample.Fcst=mf2.fcst$mean)
autoplot(window(vals, start=c(1958,1)))

coef(mf) == coef(mf2)
}
\seealso{
\code{\link[forecastR]{ts.model.fit}}
}

