% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofit.R
\name{ts.multimodel.fit}
\alias{ts.multimodel.fit}
\title{Multi- Time Series Model Fit}
\usage{
ts.multimodel.fit(y, split = 0.2, oos.h = 18L, alpha = 0.05,
  ts.model.types = package_options("autofit.models"),
  freq.multiple = package_options("short.ts.frequency.multiple"), ...)
}
\arguments{
\item{y}{\code{\link[stats]{ts}} object. univariate time series}

\item{split}{argument passed to \code{ts.split}}

\item{oos.h}{integer. Out of sample forecast steps to be provided in the
output.}

\item{alpha}{significance level for comparative testing}

\item{ts.model.types}{character vector. name of function returing \code{tsm}
model object}

\item{freq.multiple}{numeric. see \code{\link[forecastR]{short.ts.test}}}

\item{...}{extra arguments for `ts.model.type`}
}
\description{
Fits all models in `ts.model.types`, with the addition of providing a naive
forecast, a sample mean forecast and fitting a theta model forecast. Relative
Error returned is defined as `error <- estimate - actual`.
}
\examples{
library(ggplot2)
library(forecastR)
data('AirPassengers', package='datasets')
mf <- ts.multimodel.fit(AirPassengers)
vals <- cbind(mf$data, mf$train.experts)
colnames(vals) <- c("Data", colnames(mf$train.experts))

autoplot(window(vals, start=c(1958,1)))
}
\seealso{
\code{\link[forecast]{naive}} \code{\link[forecast]{meanf}}
  \code{\link[forecast]{thetaf}}
}

