% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofit.R
\name{ts.multimodel}
\alias{ts.multimodel}
\title{Multimodel fit}
\usage{
ts.multimodel(y, split = 0.2, oos.h = 18L, alpha = 0.05,
  ts.model.types = c("arima", "sarima", "arfima", "bats", "ets",
  "ets.multiplicative", "nnetar", "nnetar.w.decay", "tbats"), ...)
}
\arguments{
\item{y}{\code{\link[stats]{ts}} object. univariate time series}

\item{split}{argument passed to \code{ts.split}}

\item{oos.h}{integer. Out of sample forecast steps to be provided in the
output.}

\item{alpha}{significance level for comparative testing}

\item{ts.model.types}{character vector. name of function returing \code{tsm}
model object}

\item{...}{extra arguments for `ts.model.type`}
}
\description{
Multimodel fit
}
\examples{
library(ggplot2)
library(forecast)
library(forecastR)
data('AirPassengers', package='datasets')
mf <- ts.multimodel(AirPassengers)
autoplot(mf$data)
}

