% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodel_fit.R
\name{ts.multimodel.refit}
\alias{ts.multimodel.refit}
\title{Update Multimodel fit}
\usage{
ts.multimodel.refit(y, tsm.multi, ...)
}
\arguments{
\item{y}{\code{ts}. Univariate Time Series}

\item{tsm.multi}{\code{tsm.multi} object. Model which was fit in call to
\code{\link[forecastR]{ts.multimodel.fit}}}

\item{...}{additional keyword arguments provided to `ts.model.type` function.}
}
\value{
\code{tsm.mulit} object, with subspecialization determined by
  `ts.model.type` argument.
}
\description{
Update Multimodel fit
}
\examples{
library(ggplot2)
library(forecastR)
data('AirPassengers', package = 'datasets')

ap.split <- ts.split(AirPassengers, split = 24)
in.sample <- ap.split$in.sample

mf <- ts.multimodel.fit(in.sample)

mf.upd <- ts.multimodel.refit(ap.split$data, mf)
}
\seealso{
\code{\link[forecastR]{ts.multimodel.fit}}
}

