% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodel_fit.R
\name{ts.multimodel.resample}
\alias{ts.multimodel.resample}
\title{Update Multimodel Fit}
\usage{
ts.multimodel.resample(y, tsm.multi, boot.reps = 25, split = 0.2,
  oos.h = 18L, alpha = 0.05, ...)
}
\arguments{
\item{y}{\code{ts}. Univariate Time Series}

\item{tsm.multi}{\code{tsm.multi} object. Model which was fit in call to
\code{\link[forecastR]{ts.model.fit}}}

\item{boot.reps}{integer. Number of bootstrap replicates to test against}

\item{split}{numeric. Determines in/out-sample proportions}

\item{oos.h}{integer. Number of out of sample forecast steps}

\item{alpha}{numeric. Significance threshold for internal testing}

\item{...}{additional keyword arguments provided to \code{\link[meboot]{meboot}}
and the models when calling their update methods.}
}
\value{
\code{tsm} object, with subspecialization determined by `ts.model.type`
argument.
}
\description{
Update Multimodel Fit
}
\examples{
library(ggplot2)
library(forecastR)
data('AirPassengers', package = 'datasets')

mmf <- ts.multimodel.fit(AirPassengers)

mmr <- ts.multimodel.resample(AirPassengers, mmf, boot.reps = 25)
}
\seealso{
\code{\link[forecastR]{ts.model.fit}}
}

