% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsm.R
\name{tsm}
\alias{is.tsm}
\alias{tsm}
\title{Wrapper for Time Series Models that creates a unified interface.}
\usage{
tsm(function.name, y, ...)

is.tsm(object)
}
\arguments{
\item{function.name}{function name for the model}

\item{y}{fitted time series model}

\item{...}{not currently used}

\item{object}{value to be validated as `tsm` object}
}
\value{
object of class \code{tsm}

logical value
}
\description{
Internal function which applies \code{tsm} object formatting to time series
models.
}
\section{Functions}{
\itemize{
\item \code{is.tsm}: Inheritance check for \code{tsm} object
}}
\seealso{
\code{\link[methods]{is}}
}

