% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{arfima}
\alias{arfima}
\title{ARFIMA model fitting and updates}
\usage{
arfima(y, ..., model.name = "arfima")
}
\arguments{
\item{y}{Univariate Time Series}

\item{...}{additional argument for model function}

\item{model.name}{name of model specialization}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{arfima}} to
allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
fit <- arfima(AirPassengers)
fit2 <- arfima(AirPassengers, lambda=0)
}
\seealso{
\code{\link[forecast]{arfima}}
}

