% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{arfima}
\alias{arfima}
\title{ARFIMA model fitting and updates}
\usage{
arfima(y, lambda = NULL, model = NULL, ...)
}
\arguments{
\item{y}{Univariate Time Series}

\item{lambda}{numeric. Box-Cox transformation parameter.}

\item{model}{\code{tsm} object, used to refit model}

\item{...}{additional argument for model function}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{arfima}} to
allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
fit <- arfima(AirPassengers)
fit2 <- arfima(AirPassengers, lambda=0)
}
\seealso{
\code{\link[forecast]{arfima}}

Other Time Series Models: \code{\link{arima}},
  \code{\link{bats}}, \code{\link{ets}}
}

