% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_model_arima.R
\name{arima}
\alias{arima}
\title{ARIMA model fitting and updates}
\usage{
arima(y, lambda = NULL, model = NULL, order = NULL, ...)
}
\arguments{
\item{y}{Univariate Time Series}

\item{lambda}{optional, if null ignored otherwise treated as the Box-Cox
parameter}

\item{model}{\code{tsm}, \code{\link[forecast]{Arima}} or
\code{\link[stats]{arima}} object to be refit}

\item{order}{integer vector. Non-seasonal ARIMA order}

\item{...}{additional argument for model function. See
\code{\link[forecast]{Arima}} and \code{\link[forecast]{auto.arima}}
for allowable named arguments and definitions}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{Arima}} and
\code{\link[forecast]{auto.arima}} to allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package = 'datasets')
fit <- arima(AirPassengers)
fit2 <- arima(AirPassengers, D = 1)
fit3 <- arima(AirPassengers, lambda = 0)
}
\seealso{
\code{\link[forecast]{Arima}} and \code{\link[forecast]{auto.arima}}

Other Time Series Models: \code{\link{arfima}},
  \code{\link{bats}}, \code{\link{ets}}
}

