% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ets}
\alias{ets}
\title{ETS model fitting and updates}
\usage{
ets(y, model = "ZZZ", allow.multiplicative.trend = TRUE, ...)
}
\arguments{
\item{y}{Univariate Time Series}

\item{model}{either a string identifying model method, or previously fitted
model object. See \code{\link[forecast]{ets}} for more information}

\item{allow.multiplicative.trend}{logical. should the model consider
multiplicative exponential smoothing models?}

\item{...}{additional argument for model function}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{ets}} to
allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package = 'datasets')
fit <- ets(AirPassengers)
fit2 <- ets(AirPassengers, lambda = 0)
}
\seealso{
\code{\link[forecast]{ets}}

Other Time Series Models: \code{\link{arfima}},
  \code{\link{arima}}, \code{\link{bats}}
}

