% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.tests.R
\name{mcleod.li.test}
\alias{mcleod.li.test}
\alias{mcleod.li.test.Arima}
\alias{mcleod.li.test.default}
\alias{mcleod.li.test.ts}
\alias{mcleod.li.test.tsm}
\title{McLeod - Li Test for autocorrelation}
\usage{
mcleod.li.test(y, alpha = 0.05, max.lag = autocorr.lags(y, alpha))

\method{mcleod.li.test}{default}(y, alpha = 0.05, max.lag = autocorr.lags(y,
  alpha))

\method{mcleod.li.test}{ts}(y, alpha = 0.05, max.lag = autocorr.lags(y,
  alpha))

\method{mcleod.li.test}{tsm}(y, alpha = 0.05,
  max.lag = autocorr.lags(residuals(model(y)), alpha))

\method{mcleod.li.test}{Arima}(y, alpha = 0.05,
  max.lag = autocorr.lags(residuals(model(y)), alpha))
}
\arguments{
\item{y}{univariate time series}

\item{alpha}{numeric, determines significance threshold for p-value analysis}

\item{max.lag}{integer. If undefined, calculates an appropriate value}
}
\value{
\code{'McLeodLiTest'} object
}
\description{
McLeod - Li Test for autocorrelation
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: \code{numeric} specialization of the McLeod-Li Test.

\item \code{ts}: \code{ts} specialization of the McLeod-Li Test.

\item \code{tsm}: \code{tsm} specialization of the McLeod-Li Test.

\item \code{Arima}: \code{\link[forecast]{Arima}} specialization of the McLeod-Li Test
}}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
aa <- arima(AirPassengers)
mcleod.li.test(aa)
aa2 <- arima(AirPassengers, lambda=0)
mcleod.li.test(aa2)

#check for GARCH residuals
mcleod.li.test(residuals(aa)^2)
mcleod.li.test(residuals(aa2)^2)
}
\seealso{
\code{\link[TSA]{McLeod.Li.test}}

Other Autocorrelation Tests: \code{\link{autocorr.lags}},
  \code{\link{durbin.watson.test}},
  \code{\link{ljung.box.test}}, \code{\link{print.act}}
}

