% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsm.R
\name{model}
\alias{model}
\alias{model.default}
\title{Extract Model from \code{object}}
\usage{
model(object, ...)

\method{model}{default}(object, ...)
}
\arguments{
\item{object}{an object which stores a model which requires direct access.}

\item{...}{other arguments}
}
\description{
Extract Model from \code{object}
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default model extractor as pass-through
}}

