% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_model_nnetar.R
\name{nnetar}
\alias{nnetar}
\title{NNETAR model fitting and updates}
\usage{
nnetar(y, model = NULL, lambda = NULL, ...)
}
\arguments{
\item{y}{Univariate Time Series}

\item{model}{previously fit model object}

\item{lambda}{optional, if null ignored otherwise treated as the Box-Cox
parameter}

\item{...}{additional argument for model function}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{nnetar}} to
allow unified interface.
}
\examples{
library(ggplot2)
library(forecastR)
data('AirPassengers', package = 'datasets')
ap.split <- ts.split(AirPassengers)
out.sample.len <- length(ap.split$out.of.sample)

#base fit
fit <- nnetar(ap.split$in.sample)
#log transformation
fit2 <- nnetar(ap.split$in.sample, lambda = 0)
#log transformation w/ nnet decay parameter
fit3 <- nnetar(ap.split$in.sample, lambda = 0, decay = 0.01)

fcst <- forecast(fit, h = out.sample.len)
fcst2 <- forecast(fit2, h = out.sample.len)
fcst3 <- forecast(fit3, h = out.sample.len)

vals <- window(cbind(
 data = ap.split$data,
 base = fcst$mean,
 lambda = fcst2$mean,
 decay = fcst3$mean
 ), start = c(1957,1))

## suppressWarnings(autoplot(vals, na.rm = TRUE))
}
\seealso{
\code{\link[forecast]{nnetar}}
}

