% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{stlm}
\alias{stlm}
\title{STLM model fitting and updates}
\usage{
stlm(y, ...)
}
\arguments{
\item{y}{Univariate Time Series}

\item{...}{additional argument for model function}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{stlm}} to
allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package='datasets')
fit <- stlm(AirPassengers)
fit2 <- stlm(AirPassengers, lambda=0)
}
\seealso{
\code{\link[forecast]{tbats}}
}

