% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofit.R
\name{ts.mixture}
\alias{ts.mixture}
\title{Mixture Model fit}
\usage{
ts.mixture(y, split = 0.2, oos.h = 18L, alpha = 0.05,
  ts.model.types = c("arima", "arfima", "bats", "ets", "nnetar",
  "nnetar.w.decay", "tbats"), ..., parent = emptyenv())
}
\arguments{
\item{y}{\code{\link[stats]{ts}} object. univariate time series}

\item{split}{argument passed to \code{ts.split}}

\item{oos.h}{integer. Out of sample forecast steps to be provided in the
output.}

\item{alpha}{significance level for comparative testing}

\item{ts.model.types}{character vector. name of function returing \code{tsm}
model object}

\item{...}{extra arguments for `ts.model.type`}

\item{parent}{\code{environment} object. Default is \code{\link[base]{emptyenv}}}
}
\description{
Mixture Model fit
}

