% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model_fit.R
\name{ts.mixture}
\alias{ts.mixture}
\title{Mixture Model fit}
\usage{
ts.mixture(y, tsm.multi = NULL, split = 0.2, oos.h = 18L, alpha = 0.05,
  boot.reps = NULL, ...)
}
\arguments{
\item{y}{\code{\link[stats]{ts}} object. univariate time series}

\item{tsm.multi}{\code{tsm.multi} object returned from \code{\link[forecastR]{ts.multimodel.fit}}}

\item{split}{argument passed to \code{ts.split}}

\item{oos.h}{integer. Out of sample forecast steps to be provided in the
output.}

\item{alpha}{significance level for comparative testing}

\item{boot.reps}{integer. number of bootstrap replicates to fit against}

\item{...}{extra arguments for `ts.model.type`}
}
\description{
Mixture Model fit
}
\examples{
library(ggplot2)
library(forecastR)
library(reshape2)

data("AirPassengers", package="datasets")

split.fctr <- 24
ap <- ts.split(AirPassengers, split = split.fctr, as.list=TRUE)
ap.ts <- cbind(in.sample = ap$in.sample, out.of.sample = ap$out.of.sample)
ap.df <- as.data.frame(ap.ts)
ap.df$date <- as.Date(ISOdate(as.integer(time(ap$data)),
                              as.integer(cycle(ap$data)), 1))
ap.melt.df <- melt(ap.df, id.vars = "date", na.rm = TRUE)

(ggplot(ap.melt.df, aes(x = date, y = value, color = variable)) +
  geom_line() +
  ggtitle("Test/Training Set View"))

tsm <- ts.multimodel.fit(AirPassengers, split=split.fctr)
mx <- ts.mixture(AirPassengers, tsm, split=split.fctr, oos.h = 12)

vals <- cbind(AirPassengers, mx$mixture.forecast, mx$forecast)
vals <- window(vals, start=1959)
vals.df <- as.data.frame(vals)
colnames(vals.df) <- c("Data", "Mixture Model", as.character(colnames(mx$forecast)))
vals.df$date <- as.Date(ISOdate(as.integer(time(vals)),
                                as.integer(cycle(vals)), 1))
vals.melt <- melt(vals.df, id.var = "date", na.rm = TRUE)
(ggplot(vals.melt, aes(x = date, y = value, color = variable)) +
  geom_line() +
  ggtitle("All Forecasts"))

fcst <- cbind(mx$mixture.forecast, mx$test.forecast)
fcst.err <- fcst - ap$out.of.sample
colnames(fcst.err) <- c("Mixture Model", colnames(mx$test.forecast))
fcst.err.df <- as.data.frame(fcst.err)
fcst.err.df$date <- as.Date(ISOdate(as.integer(time(fcst.err)),
                                    as.integer(cycle(fcst.err)), 1))
fcst.err.melt.df <- melt(fcst.err.df, id.var = "date", na.rm = TRUE)
(ggplot(fcst.err.melt.df, aes(x = date, y = value, color = variable)) +
  geom_line() +
  ggtitle("Error"))

fcst.err.melt.df$value <- abs(fcst.err.melt.df$value)
(ggplot(fcst.err.melt.df, aes(x = date, y = value, color = variable)) +
  geom_line() +
  ggtitle("Absolute Error"))

}

