% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofit.R
\name{ts.model.fit}
\alias{ts.model.fit}
\title{Generic Time Series Model Fit for Package Models}
\usage{
ts.model.fit(y, ts.model.type = package_options("autofit.models"), ...)
}
\arguments{
\item{y}{\code{ts}. Univariate Time Series}

\item{ts.model.type}{character. Name of model function from package.}

\item{...}{additional keyword arguments provided to `ts.model.type` function.}
}
\value{
\code{tsm} object, with subspecialization determined by `ts.model.type`
argument.
}
\description{
Generic Time Series Model Fit for Package Models
}
\examples{
library(ggplot2)
library(forecastR)
data('AirPassengers', package = 'datasets')

ap.split <- ts.split(AirPassengers)

mf <- ts.model.fit(ap.split$in.sample, ts.model.type = 'arima')
mf.fcst <- forecast(mf, length(ap.split$out.of.sample))
summary(mf)
coef(mf)

autoplot(resid(mf))

vals <- cbind(Data = ap.split$data, Forecast.Mean = mf.fcst$mean)
autoplot(window(vals, start = c(1958,1)), na.rm=TRUE)
}
\seealso{
\code{\link[forecastR]{arima}} \code{\link[forecastR]{arfima}}
  \code{\link[forecastR]{bats}} \code{\link[forecastR]{ets}}
  \code{\link[forecastR]{nnetar}} \code{\link[forecastR]{tbats}}
  \code{\link[forecast]{auto.arima}} \code{\link[forecast]{Arima}}
  \code{\link[forecast]{arfima}} \code{\link[forecast]{bats}}
  \code{\link[forecast]{ets}} \code{\link[forecast]{nnetar}}
  \code{\link[forecast]{tbats}}
}

