% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofit.R
\name{ts.multimodel.refit}
\alias{ts.multimodel.refit}
\title{Update Multimodel fit}
\usage{
ts.multimodel.refit(y, model.list, split = 0.2, oos.h = 18L, alpha = 0.05,
  return.models = TRUE, ...)
}
\arguments{
\item{y}{\code{ts}. Univariate Time Series}

\item{model.list}{\code{tsm.multi} object. Model which was fit in call to
\code{\link[forecastR]{ts.multimodel.fit}}}

\item{split}{argument passed to \code{ts.split}}

\item{oos.h}{integer. Out of sample forecast steps to be provided in the
output.}

\item{alpha}{significance level for comparative testing}

\item{return.models}{logical. If \code{TRUE}, models are returned in the
output list as `models` argument, otherwise does not return models.}

\item{...}{additional keyword arguments provided to `ts.model.type` function.}
}
\value{
\code{tsm.mulit} object, with subspecialization determined by
  `ts.model.type` argument.
}
\description{
Update Multimodel fit
}
\examples{
library(ggplot2)
library(forecastR)
data('AirPassengers', package = 'datasets')

ap.split <- ts.split(AirPassengers)

mf <- ts.model.fit(ap.split$in.sample, ts.model.type = 'arima')
mf.fcst <- forecast(mf, length(ap.split$out.of.sample))
summary(mf)

vals <- cbind(Data = ap.split$data, Forecast.Mean = mf.fcst$mean)
autoplot(window(vals, start = c(1958,1)), na.rm=TRUE)

mf2 <- ts.model.refit(ap.split$data, mf)
mf2.fcst <- forecast(mf2, 24)
summary(mf2)

vals <- cbind(Data = ap.split$data, In.Sample.Fcst = mf.fcst$mean,
              Out.Of.Sample.Fcst = mf2.fcst$mean)
autoplot(window(vals, start = c(1958,1)), na.rm=TRUE)

coef(mf) ==  coef(mf2)
}
\seealso{
\code{\link[forecastR]{ts.multimodel.fit}}
}

