% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{tslm}
\alias{tslm}
\title{Time Series Linear Model}
\usage{
tslm(y, model = NULL, lambda = NULL, ...)
}
\arguments{
\item{y}{Univariate Time Series}

\item{model}{previously fit model}

\item{lambda}{optional, if null ignored otherwise treated as the Box-Cox
parameter}

\item{...}{additional arguments to \code{\link[forecast]{tslm}}}
}
\value{
\code{tsm} object
}
\description{
Wrapper for forecast package functions \code{\link[forecast]{tslm}} to
allow unified interface.
}
\examples{
library(forecastR)
data('AirPassengers', package = 'datasets')
fit <- nnetar(AirPassengers)
fit2 <- nnetar(AirPassengers, lambda = 0)
}
\seealso{
\code{\link[forecast]{tslm}}
}

