% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsm.R
\name{tsm}
\alias{tsm}
\title{Wrapper for Time Series Models that creates a unified interface.}
\usage{
tsm(function.name, y, ...)
}
\arguments{
\item{function.name}{function name for the model}

\item{y}{fitted time series model}

\item{...}{not currently used}
}
\value{
object of class \code{tsm}
}
\description{
Internal function which applies \code{tsm} object formatting to time series
models.
}

